% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_trial_data.R
\name{create_new_trial_data}
\alias{create_new_trial_data}
\title{Data on new trial in target population}
\usage{
create_new_trial_data(n_total, est, se)
}
\arguments{
\item{n_total}{The total sample size.}

\item{est}{Treatment effect estimate.}

\item{se}{Standard error of the treatment effect estimate.}
}
\value{
A numeric vector with data on the new trial, incl. quantiles of an assumed normal data likelihood.
}
\description{
Creates a vector containing data on the new trial in the target population. This may be hypothetical data in the planning stage.
}
\examples{
new_trial_data <- create_new_trial_data(
  n_total = 30, est = 1.27, se = 0.95
)

}
\seealso{
\code{\link{create_posterior_data}}, \code{\link{create_tipmap_data}}
}
