% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{default_quantiles}
\alias{default_quantiles}
\title{Default quantiles}
\format{
An object of class \code{numeric} of length 13.
}
\usage{
default_quantiles
}
\description{
Default quantiles
}
\keyword{datasets}
