% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{default_weights}
\alias{default_weights}
\title{Default weights}
\format{
An object of class \code{numeric} of length 201.
}
\usage{
default_weights
}
\description{
Default weights
}
\keyword{datasets}
