% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{defaultWeights}
\alias{defaultWeights}
\title{Default weights}
\format{
An object of class \code{numeric} of length 201.
}
\usage{
defaultWeights
}
\description{
Default weights
}
\keyword{datasets}
