% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_1exp.R
\name{fit_beta_1exp}
\alias{fit_beta_1exp}
\title{Fit beta distribution for one expert}
\usage{
fit_beta_1exp(df)
}
\arguments{
\item{df}{A dataframe generated by \code{get_model_input_1exp}.}
}
\value{
Parameters of beta fit.
}
\description{
Fit beta distribution to weights elicited from one expert via the roulette method.
}
\details{
This function is based on \code{SHELF::fitdist} and yields identical results.
}
\examples{
chips <- c(0, 2, 3, 2, 1, 1, 1, 0, 0, 0)
x <- get_cum_probs_1exp(chips)
y <- get_model_input_1exp(x)
fit_beta_1exp(df = y)["par"]

}
