% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_mult_exp.R
\name{fit_beta_mult_exp}
\alias{fit_beta_mult_exp}
\title{Fit beta distributions for multiple experts}
\usage{
fit_beta_mult_exp(chips_mult)
}
\arguments{
\item{chips_mult}{A dataframe or matrix containing weights.
It should contain one row per expert and 10 columns, one for each bin, representing weights from 0 to 1.}
}
\value{
Parameters of the individual beta distributions.
}
\description{
Fit beta distributions to weights elicited from multiple experts via the roulette method.
}
\examples{
beta_fits <- fit_beta_mult_exp(
chips_mult =
  rbind(
      c(0, 0, 0, 0, 2, 3, 3, 2, 0, 0),
      c(0, 0, 0, 1, 2, 4, 2, 1, 0, 0),
      c(0, 0, 0, 2, 2, 2, 2, 2, 0, 0)
      )
  )
beta_fits

}
