% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cum_probs_1exp.R
\name{get_cum_probs_1exp}
\alias{get_cum_probs_1exp}
\title{Get cumulative probabilities from distribution of chips of one expert}
\usage{
get_cum_probs_1exp(chips)
}
\arguments{
\item{chips}{Vector of integers, representing the distribution of chips assigned by one expert, as elicited through the roulette method.
Each element of the vector represents one bin in the grid.}
}
\value{
A numeric vector with the cumulative distribution of chips.
}
\description{
Get cumulative probabilities from distribution of chips of one expert
}
\examples{
chips <- c(0, 2, 3, 2, 1, 1, 1, 0, 0, 0)
x <- get_cum_probs_1exp(chips)
print(x)

}
\seealso{
\code{\link{get_model_input_1exp}} and \code{\link{fit_beta_1exp}}.
}
