% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_by_weight.R
\name{get_posterior_by_weight}
\alias{get_posterior_by_weight}
\title{Filter posterior by given weights}
\usage{
get_posterior_by_weight(posterior, weight)
}
\arguments{
\item{posterior}{The posterior data to be filtered (see \code{create_posterior_data()}).}

\item{weight}{The weight(s) to be filtered by.}
}
\value{
The filtered posterior values
}
\description{
Returns quantiles of the posterior distribution of the treatment effect for one or more specified weights.
}
\examples{
get_posterior_by_weight(
  posterior = load_tipmap_data("tipPost.rds"),
  weight = c(0.05, 0.1)
)

}
\seealso{
\code{\link{create_posterior_data}}
}
