% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_mult_exp.R
\name{get_summary_mult_exp}
\alias{get_summary_mult_exp}
\title{Summarize expert weights}
\usage{
get_summary_mult_exp(chips_mult, n = 500, weights = NULL)
}
\arguments{
\item{chips_mult}{A data frame or matrix containing expert weights.}

\item{n}{The number of samples to be drawn to obtain summary statistics. Defaults to 500.}

\item{weights}{Weights assigned to each expert. Defaults to equal weights.}
}
\value{
A vector containing summary statistics.
}
\description{
Computes minimum, maximum, mean and quartiles for expert weights.
}
\examples{
get_summary_mult_exp(rbind(
  c(0, 0, 0, 0, 2, 3, 3, 2, 0, 0),
  c(0, 0, 0, 1, 2, 4, 2, 1, 0, 0),
  c(0, 0, 0, 2, 2, 2, 2, 2, 0, 0)
  ))

}
