% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitationFunctions.R
\name{getExpertCumulativeProbs}
\alias{getExpertCumulativeProbs}
\title{Get cumulative probabilities for specified chip distribution}
\usage{
getExpertCumulativeProbs(chips)
}
\arguments{
\item{chips}{A numeric vector representing chip distribution for an expert.
Vector must be of length 10 and contents must add up to 10. First column represents weight 0-0.1.}
}
\value{
The cumulative probabilities for the expert.
}
\description{
Internal function needed for expert elicitation methods.
}
\examples{
getExpertCumulativeProbs(c(0, 0, 0, 2, 2, 2, 2, 2, 0, 0))

}
