% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitationFunctions.R
\name{getModelInputData}
\alias{getModelInputData}
\title{Transform cumulative probabilities to fit beta distributions.}
\usage{
getModelInputData(cumprobs, w)
}
\arguments{
\item{cumprobs}{Cumulative probabilities for a single expert. See \code{getExpertCumulativeProbs()}.}

\item{w}{Weight of bins.}
}
\value{
Data to fit beta distributions.
}
\description{
Internal function needed for expert elicitation methods.
}
