% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosteriorByWeight.R
\name{getPosteriorByWeight}
\alias{getPosteriorByWeight}
\title{Filter posterior by given weights}
\usage{
getPosteriorByWeight(posterior, weight)
}
\arguments{
\item{posterior}{The posterior to be filtered (see \code{createPosteriorData()}).}

\item{weight}{The weight(s) to be filtered by.}
}
\value{
The filtered posterior values
}
\description{
Returns quantiles for the given posterior filtered by specified weights.
}
\examples{
getPosteriorByWeight(
  posterior = loadTipmapData("tipPost.rds"),
  weight = c(0.05, 0.1)
)

}
\seealso{
\code{\link{createPosteriorData}}
}
