% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tipmap_data.R
\name{load_tipmap_data}
\alias{load_tipmap_data}
\title{Load exemplary datasets}
\usage{
load_tipmap_data(file)
}
\arguments{
\item{file}{The dataset to be loaded.}
}
\value{
A pre-saved dataset.
}
\description{
Loads one of three exemplary datasets in the package.
}
\examples{

load_tipmap_data(file = "tipdat.rds")
load_tipmap_data(file = "tipmapPrior.rds")
load_tipmap_data(file = "tipPost.rds")

}
