% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTipmapData.R
\name{loadTipmapData}
\alias{loadTipmapData}
\title{Load pre-specified data sets for examples}
\usage{
loadTipmapData(file)
}
\arguments{
\item{file}{The dataset to be loaded}
}
\value{
A pre-saved dataset
}
\description{
Loads one of three pre-specified data sets ready to be used in functions
}
\examples{

loadTipmapData(file = "tipdat.rds")

loadTipmapData(file = "tipmapPrior.rds")

loadTipmapData(file = "tipPost.rds")

}
