% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_bias.R
\name{oc_bias}
\alias{oc_bias}
\title{Assessing bias}
\usage{
oc_bias(m, se, true_effect, weights, map_prior, sigma)
}
\arguments{
\item{m}{Numerical vector of simulated effect estimates.}

\item{se}{Numerical vector of simulated standard errors (\code{m} and \code{se} need to have the same length).}

\item{true_effect}{Numerical value, representing the true treatment effect (usually the mean of the simulated \code{m}).}

\item{weights}{Vector of weights of the informative component of the MAP prior.}

\item{map_prior}{A MAP prior containing information about the trial(s) in the source population, created using \code{RBesT}.}

\item{sigma}{Standard deviation of the weakly informative component of the MAP prior, recommended to be the unit-information standard deviation.}
}
\value{
A 2-dimensional array containing results on bias.
}
\description{
Assessment of absolute bias in posterior means and medians for a given weight and evidence level, using simulated data as input.
}
\examples{
set.seed(123)
n_sims <- 10 # small number for exemplary application 
sim_dat <- list(
  "m" = rnorm(n = n_sims, mean = 1.15, sd = 0.1),
  "se" = rnorm(n = n_sims, mean = 1.8, sd = 0.3)
)
results <- oc_bias(
  m = sim_dat[["m"]],
  se = sim_dat[["se"]],
  true_effect = 1.15,
  weights = seq(0, 1, by = 0.01), 
  map_prior = load_tipmap_data("tipmapPrior.rds"), 
  sigma = 16.23
) 
print(results)
}
\seealso{
\code{\link{oc_pos}} and \code{\link{oc_coverage}}
}
