% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_pos.R
\name{oc_pos}
\alias{oc_pos}
\title{Probability of success calculation}
\usage{
oc_pos(
  m,
  se,
  probs,
  weights,
  map_prior,
  sigma,
  null_effect = 0,
  direction_pos = TRUE
)
}
\arguments{
\item{m}{Numerical vector of simulated effect estimates.}

\item{se}{Numerical vector of simulated standard errors (`m` and `se` need to have the same length).}

\item{probs}{Vector of quantiles q, with 1 minus q representing an evidence level of interest (where positive effect estimate indicate a beneficial treatment).}

\item{weights}{Vector of weights of the informative component of the MAP prior.}

\item{map_prior}{A MAP prior containing information about the trial(s) in the source population, created using \code{RBesT}.}

\item{sigma}{Standard deviation to be used for the weakly informative component of the MAP prior, recommended to be the unit-information standard deviation.}

\item{null_effect}{Numerical value representing the null effect (defaults to 0).}

\item{direction_pos}{Logical value, `TRUE` (default) if effects greater that the `null_effect` indicate a beneficial treatment and `FALSE` otherwise.}
}
\value{
An array of probabilities, either of truly (probability of success) or falsely rejecting the null hypothesis of interest for a given weight and evidence level.
}
\description{
Calculate the probability of truly (probability of success) or falsely rejecting the null hypothesis of interest for a given weight and evidence level.
}
\examples{
set.seed(123)
n_sims <- 5 # small number for exemplary application
sim_dat <- list(
  "m" = rnorm(n = n_sims, mean = 1.15, sd = 0.1),
  "se" = rnorm(n = n_sims, mean = 1.8, sd = 0.3)
)
pos_dat <- oc_pos(
  m = sim_dat[["m"]],
  se = sim_dat[["se"]],
  probs = c(0.025, 0.05, 0.1, 0.2), 
  weights = seq(0, 1, by = 0.01), 
  map_prior = load_tipmap_data("tipmapPrior.rds"), 
  sigma = 16.23,
  null_effect = 0,
  direction_pos = TRUE
) 
print(pos_dat)
}
