% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitationFunctions.R
\name{rmixbeta}
\alias{rmixbeta}
\title{Draw a single sample from a mixture of specified beta distributions.}
\usage{
rmixbeta(betaParameters, weights = NULL)
}
\arguments{
\item{betaParameters}{Parameters for each beta distribution.}

\item{weights}{Optional vector of weights assigned to experts. Defaults to uniform.}
}
\value{
A random sample from a mixture of specified beta distributions.
}
\description{
Internal function needed for expert elicitation methods.
}
\seealso{
\code{\link{fitExpertBetaDistributions}}
}
