% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipmap-package.R
\docType{package}
\name{tipmap-package}
\alias{tipmap}
\alias{tipmap-package}
\title{tipmap: Tipping Point Analysis for Bayesian Dynamic Borrowing}
\description{
Tipping point analysis for clinical trials that employ Bayesian dynamic borrowing via robust meta-analytic predictive (MAP) priors. Further functions facilitate expert elicitation of a primary weight of the informative component of the robust MAP prior and computation of operating characteristics. Intended use is the planning, analysis and interpretation of extrapolation studies in pediatric drug development, but applicability is generally wider.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Boehringer-Ingelheim/tipmap}
  \item Report bugs at \url{https://github.com/Boehringer-Ingelheim/tipmap/issues}
}

}
\author{
\strong{Maintainer}: Christian Stock \email{christian.stock@boehringer-ingelheim.com} (\href{https://orcid.org/0000-0002-3493-3234}{ORCID})

Authors:
\itemize{
  \item Morten Dreher
}

Other contributors:
\itemize{
  \item Emma Torrini [contributor]
  \item  Boehringer Ingelheim Pharma GmbH & Co. KG [copyright holder, funder]
}

}
\keyword{internal}
