% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{tipmap_lightred}
\alias{tipmap_lightred}
\title{Custom light red}
\format{
An object of class \code{character} of length 1.
}
\usage{
tipmap_lightred
}
\description{
Custom light red
}
\keyword{datasets}
