% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{tipmapLightRed}
\alias{tipmapLightRed}
\title{custom light red}
\format{
An object of class \code{character} of length 1.
}
\usage{
tipmapLightRed
}
\description{
custom light red
}
\keyword{datasets}
