% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{defaultQuantiles}
\alias{defaultQuantiles}
\title{default quantiles}
\format{
An object of class \code{numeric} of length 13.
}
\usage{
defaultQuantiles
}
\description{
default quantiles
}
\keyword{datasets}
