% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cum_probs_1exp.R
\name{get_cum_probs_1exp}
\alias{get_cum_probs_1exp}
\title{Get cumulative probabilities from distribution of chips of one expert}
\usage{
get_cum_probs_1exp(chips)
}
\arguments{
\item{chips}{A numeric vector representing the distribution of chips of one expert.
The vector must be of length 10 and contents must add up to 10.
The first column represents weight on interval 0 to 0.1.}
}
\value{
A vector of cumulative probabilities.
}
\description{
Internal function needed for expert elicitation methods.
}
\examples{
chips <- c(0,2,3,2,1,1,1,0,0,0)
x <- get_cum_probs_1exp(chips)
x

}
