% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_input_1exp.R
\name{get_model_input_1exp}
\alias{get_model_input_1exp}
\title{Transform cumulative probabilities to fit beta distributions}
\usage{
get_model_input_1exp(cum_probs, w = NULL)
}
\arguments{
\item{cum_probs}{Cumulative probabilities of weights of one expert.}

\item{w}{Weight of bins.}
}
\value{
Dataframe that can be used as input to fit beta distributions by
\code{fit_beta_1exp()}.
}
\description{
Internal function needed for expert elicitation methods.
}
\examples{
chips <- c(0,2,3,2,1,1,1,0,0,0)
x <- get_cum_probs_1exp(chips)
y <- get_model_input_1exp(x)
y

}
