% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipmap-package.R
\docType{package}
\name{tipmap-package}
\alias{tipmap}
\alias{tipmap-package}
\title{tipmap: Tipping Point Analysis for Bayesian Dynamic Borrowing}
\description{
Tipping point analysis for clinical trials that employ Bayesian dynamic borrowing via robust meta-analytic predictive (MAP) priors. Mainly an implementation of an approach proposed by Best and colleagues (2021) is provided \doi{10.1002/pst.2093}. Further functions facilitate the specification of the robust MAP prior via expert elicitation (using the roulette method) and computation of the posterior distribution of the treatment effect with either fixed or stochastic expert-elicited weights. Intended use is the planning, analysis and interpretation of extrapolation studies in pediatric drug development, but applicability is generally wider.
}
\author{
\strong{Maintainer}: Christian Stock \email{christian.stock@boehringer-ingelheim.com} (\href{https://orcid.org/0000-0002-3493-3234}{ORCID})

Authors:
\itemize{
  \item Morten Dreher
}

Other contributors:
\itemize{
  \item Emma Torrini [contributor]
}

}
\keyword{internal}
