% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{tipmapDarkBlue}
\alias{tipmapDarkBlue}
\title{Custom dark blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
tipmapDarkBlue
}
\description{
Custom dark blue
}
\keyword{datasets}
